+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/modules/notifications/__init__.py
@@ -1,4 +1,4 @@
-from ._notifier import publish_disk_usage
+from ._notifications_system_monitor import publish_disk_usage
 from ._setup import setup_notifications
 
 __all__: tuple[str, ...] = (
